% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsestimates.R
\name{bsestimates}
\alias{bsestimates}
\title{Bootstrap properties of estimators}
\usage{
bsestimates(data, n, sim)
}
\arguments{
\item{data}{Required Data Frame with at least 2 variables}

\item{n}{Sample size}

\item{sim}{No of simulated samples}
}
\value{
The bias, mean squared error, relative efficiency and Percentage contribution of bias
}
\description{
Bootstrap properties of estimators
}
\examples{
bsestimates(data.frame(c(2,3,1,2,3,4,5,5,6,6), c(1,1,2,3,4,4,5,3,3,5)),5, 100)
}
